+++ TinkoffASDKUI/TinkoffASDKUI/YandexPay/Interfaces/IYandexPayPaymentSheetOutput.swift
@@ -1,5 +1,5 @@
 //
-//  IYandexPayPaymentActivityOutput.swift
+//  IYandexPayPaymentSheetOutput.swift
 //  TinkoffASDKUI
 //
 //  Created by r.akhmadeev on 18.12.2022.
@@ -7,6 +7,6 @@
 
 import Foundation
 
-protocol IYandexPayPaymentActivityOutput: AnyObject {
+protocol IYandexPayPaymentSheetOutput: AnyObject {
     func yandexPayPaymentActivity(completedWith result: YandexPayPaymentResult)
 }